Adding database query manager. Refactoring component and feature
