--- components/org.wso2.carbon.analytics-common.data-provider/src/main/java/org/wso2/carbon/analytics/common/data/provider/exception/DataProviderException.java
@@ -1,44 +1,0 @@
-/*
- * Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- * WSO2 Inc. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.wso2.carbon.analytics.common.data.provider.exception;
-
-/**
- * Data Provider exception.
- */
-public class DataProviderException extends Exception {
-    private String errorMessage;
-
-    public DataProviderException(String message) {
-        super(message);
-        errorMessage = message;
-    }
-
-    public DataProviderException(String message, Throwable cause) {
-        super(message, cause);
-        errorMessage = message;
-    }
-
-    public DataProviderException(Throwable cause) {
-        super(cause);
-    }
-
-    public String getErrorMessage() {
-        return errorMessage;
-    }
-}
