+++ components/org.wso2.carbon.data.provider/pom.xml
@@ -1,0 +1,117 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<!--
+  ~ Copyright (c) 2015, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+  ~
+  ~ WSO2 Inc. licenses this file to you under the Apache License,
+  ~ Version 2.0 (the "License"); you may not use this file except
+  ~ in compliance with the License.
+  ~ You may obtain a copy of the License at
+  ~
+  ~     http://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~ Unless required by applicable law or agreed to in writing,
+  ~ software distributed under the License is distributed on an
+  ~ "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+  ~ KIND, either express or implied.  See the License for the
+  ~ specific language governing permissions and limitations
+  ~ under the License.
+  -->
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+    <parent>
+        <groupId>org.wso2.carbon.analytics-common</groupId>
+        <artifactId>org.wso2.carbon.analytics-common.parent</artifactId>
+        <version>6.0.11-SNAPSHOT</version>
+        <relativePath>../../pom.xml</relativePath>
+    </parent>
+
+    <modelVersion>4.0.0</modelVersion>
+    <artifactId>org.wso2.carbon.data.provider</artifactId>
+    <packaging>bundle</packaging>
+    <name>WSO2 Carbon - Data Provider Module</name>
+    <description>
+        Used to provide data for different clients from different sources according to the requirement
+    </description>
+    <url>http://wso2.org</url>
+
+    <dependencies>
+        <!--OSGi-->
+        <dependency>
+            <groupId>org.osgi</groupId>
+            <artifactId>org.osgi.core</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.wso2.eclipse.osgi</groupId>
+            <artifactId>org.eclipse.osgi.services</artifactId>
+        </dependency>
+
+        <!--Logging-->
+        <dependency>
+            <groupId>org.slf4j</groupId>
+            <artifactId>slf4j-api</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.slf4j</groupId>
+            <artifactId>slf4j-log4j12</artifactId>
+        </dependency>
+
+        <!--MSF4J-->
+        <dependency>
+            <groupId>org.wso2.msf4j</groupId>
+            <artifactId>msf4j-core</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>javax.websocket</groupId>
+            <artifactId>javax.websocket-api</artifactId>
+        </dependency>
+
+        <dependency>
+        <groupId>com.google.code.gson</groupId>
+            <artifactId>gson</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.wso2.carbon.datasources</groupId>
+            <artifactId>org.wso2.carbon.datasource.core</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.wso2.carbon.config</groupId>
+            <artifactId>org.wso2.carbon.config</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.wso2.carbon.analytics-common</groupId>
+            <artifactId>org.wso2.carbon.database.query.manager</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.yaml</groupId>
+            <artifactId>snakeyaml</artifactId>
+            <optional>true</optional>
+        </dependency>
+    </dependencies>
+
+    <properties>
+        <private.package>
+            org.wso2.carbon.database.query.manager.*,
+            org.wso2.carbon.database.query.manager.exception.*
+        </private.package>
+        <export.package>
+            org.wso2.carbon.data.provider.*
+        </export.package>
+        <import.package>
+            org.yaml.snakeyaml.*;version="${org.snakeyaml.import.version.range}",
+            org.wso2.carbon.database.query.manager.*,
+            org.wso2.carbon.database.query.manager.config.*,
+            org.wso2.carbon.database.query.manager.exception.*,
+            org.osgi.framework.*;version="${osgi.framework.import.version.range}",
+            org.wso2.carbon.datasource.core.*;version="${org.wso2.carbon.datasource.version.range}",
+            org.wso2.carbon.config.provider.*;version="${carbon.config.version.range}",
+            org.wso2.msf4j.*;version="${msf4j.import.version.range}",
+            javax.ws.rs.*;version="${javax.ws.rs.version.range}",
+            javax.websocket.*;version="${javax.websocket.version.range}",
+            org.slf4j.*;version="${slf4j.version.range}",
+            com.google.gson.*; version="${gson.version.range}"
+        </import.package>
+        <carbon.component>
+            osgi.service; objectClass="org.wso2.msf4j.websocket.WebSocketEndpoint"
+        </carbon.component>
+    </properties>
+
+</project>
