+++ components/org.wso2.carbon.data.provider/src/main/resources/queries.yaml
@@ -1,0 +1,25 @@
+queries:
+  -
+    mappings:
+      record_delete: ~
+      total_record_count: ~
+      record_limit: ~
+      record_greater_than: ~
+    type: default
+    version: default
+  -
+    mappings:
+      record_delete: "DELETE FROM {{TABLE_NAME}} ORDER BY {{INCREMENTAL_COLUMN}} ASC LIMIT {{LIMIT_VALUE}}"
+      total_record_count: "SELECT COUNT(*) FROM {{TABLE_NAME}}"
+      record_limit: " ORDER BY {{INCREMENTAL_COLUMN}} ASC LIMIT {{LIMIT_VALUE}}"
+      record_greater_than: " WHERE {{INCREMENTAL_COLUMN}} > {{LAST_RECORD_VALUE}}"
+    type: H2
+    version: default
+  -
+    mappings:
+      record_delete: "DELETE FROM {{TABLE_NAME}} ORDER BY {{INCREMENTAL_COLUMN}} ASC LIMIT {{LIMIT_VALUE}}"
+      total_record_count: "SELECT COUNT(*) FROM {{TABLE_NAME}}"
+      record_limit: " ORDER BY {{INCREMENTAL_COLUMN}} ASC LIMIT {{LIMIT_VALUE}}"
+      record_greater_than: " WHERE {{INCREMENTAL_COLUMN}} > {{LAST_RECORD_VALUE}}"
+    type: MySQL
+    version: default
