+++ components/org.wso2.carbon.database.query.manager/pom.xml
@@ -21,7 +21,7 @@
     <parent>
         <groupId>org.wso2.carbon.analytics-common</groupId>
         <artifactId>org.wso2.carbon.analytics-common.parent</artifactId>
-        <version>6.0.10-SNAPSHOT</version>
+        <version>6.0.11-SNAPSHOT</version>
         <relativePath>../../pom.xml</relativePath>
     </parent>
